+++ app/javascript/react/images/hero1.jpg
