+++ app/javascript/react/images/hero2.jpg
