+++ app/javascript/react/images/ianthompson.jpg
