+++ app/javascript/react/pages/jobs/(DO NOT USE)ShowCurrentJob.js
